+++ DataLayers/Framework/ModIntegrations.cs
@@ -4,7 +4,6 @@ using Pathoschild.Stardew.Common.Integrations.BetterSprinklers;
 using Pathoschild.Stardew.Common.Integrations.LineSprinklers;
 using Pathoschild.Stardew.Common.Integrations.MultiFertilizer;
 using Pathoschild.Stardew.Common.Integrations.PelicanFiber;
-using Pathoschild.Stardew.Common.Integrations.PrismaticTools;
 using Pathoschild.Stardew.Common.Integrations.SimpleSprinkler;
 using StardewModdingAPI;
 
@@ -34,9 +33,6 @@ namespace Pathoschild.Stardew.DataLayers.Framework
         /// <summary>Handles access to the Pelican Fiber mod.</summary>
         public PelicanFiberIntegration PelicanFiber { get; }
 
-        /// <summary>Handles access to the Prismatic Tools mod.</summary>
-        public PrismaticToolsIntegration PrismaticTools { get; }
-
         /// <summary>Handles access to the Simple Sprinkler mod.</summary>
         public SimpleSprinklerIntegration SimpleSprinkler { get; }
 
@@ -56,7 +52,6 @@ namespace Pathoschild.Stardew.DataLayers.Framework
             this.LineSprinklers = new LineSprinklersIntegration(modRegistry, monitor);
             this.MultiFertilizer = new MultiFertilizerIntegration(modRegistry, monitor);
             this.PelicanFiber = new PelicanFiberIntegration(modRegistry, reflection, monitor);
-            this.PrismaticTools = new PrismaticToolsIntegration(modRegistry, monitor);
             this.SimpleSprinkler = new SimpleSprinklerIntegration(modRegistry, monitor);
         }
     }