+++ SmallBeachFarm/Framework/Config/ModConfig.cs
@@ -1,5 +1,3 @@
-using StardewValley;
-
 namespace Pathoschild.Stardew.SmallBeachFarm.Framework.Config
 {
     /// <summary>The mod configuration.</summary>
@@ -13,8 +11,5 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Framework.Config
 
         /// <summary>Use the beach's background music (i.e. wave sounds) on the beach farm.</summary>
         public bool UseBeachMusic { get; set; } = false;
-
-        /// <summary>The farm layout ID to replace.</summary>
-        public int ReplaceFarmID { get; set; } = Farm.beach_layout;
     }
 }