+++ SmallBeachFarm/Patches/FarmPatcher.cs
@@ -1,13 +1,12 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using HarmonyLib;
 using Microsoft.Xna.Framework;
 using Pathoschild.Stardew.Common.Patching;
 using Pathoschild.Stardew.SmallBeachFarm.Framework;
+using Pathoschild.Stardew.SmallBeachFarm.Framework.Config;
 using StardewModdingAPI;
 using StardewValley;
-using StardewValley.Locations;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.SmallBeachFarm.Patches
@@ -21,11 +20,8 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
         /// <summary>Encapsulates logging for the Harmony patch.</summary>
         private static IMonitor Monitor;
 
-        /// <summary>Use the beach's background music (i.e. wave sounds) on the beach farm.</summary>
-        private static bool UseBeachMusic;
-
-        /// <summary>Whether to add the campfire to the farm map.</summary>
-        private static bool AddCampfire;
+        /// <summary>The mod configuration.</summary>
+        private static ModConfig Config;
 
         /// <summary>Get whether the given location is the Small Beach Farm.</summary>
         private static Func<GameLocation, bool> IsSmallBeachFarm;
@@ -36,31 +32,19 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
         /// <summary>Whether the mod is currently applying patch changes (to avoid infinite recursion),</summary>
         private static bool IsInPatch;
 
-        /// <summary>The arrival tiles by farm type when the player warps from the <see cref="IslandWest"/> obelisk, with a <c>-1</c> key for the default.</summary>
-        /// <remarks>Derived from <see cref="IslandWest.performAction"/>.</remarks>
-        [Obsolete("This will no longer be needed in Stardew Valley 1.5.5.")]
-        private static readonly Dictionary<int, Point> IslandWarpTargets = new()
-        {
-            [Farm.fourCorners_layout] = new Point(48, 39),
-            [Farm.beach_layout] = new Point(81, 29),
-            [-1] = new Point(48, 7)
-        };
-
 
         /*********
         ** Public methods
         *********/
         /// <summary>Initialize the patcher.</summary>
         /// <param name="monitor">Encapsulates logging for the Harmony patch.</param>
-        /// <param name="addCampfire">Whether to add the campfire to the farm map.</param>
-        /// <param name="useBeachMusic">Use the beach's background music (i.e. wave sounds) on the beach farm.</param>
+        /// <param name="config">The mod configuration.</param>
         /// <param name="isSmallBeachFarm">Get whether the given location is the Small Beach Farm.</param>
         /// <param name="getFishType">Get the fish that should be available from the given tile.</param>
-        public FarmPatcher(IMonitor monitor, bool addCampfire, bool useBeachMusic, Func<GameLocation, bool> isSmallBeachFarm, Func<Farm, int, int, FishType> getFishType)
+        public FarmPatcher(IMonitor monitor, ModConfig config, Func<GameLocation, bool> isSmallBeachFarm, Func<Farm, int, int, FishType> getFishType)
         {
             FarmPatcher.Monitor = monitor;
-            FarmPatcher.AddCampfire = addCampfire;
-            FarmPatcher.UseBeachMusic = useBeachMusic;
+            FarmPatcher.Config = config;
             FarmPatcher.IsSmallBeachFarm = isSmallBeachFarm;
             FarmPatcher.GetFishType = getFishType;
         }
@@ -83,6 +67,10 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
             harmony.Patch(
                 original: this.RequireMethod<GameLocation>(nameof(GameLocation.cleanupBeforePlayerExit)),
                 prefix: this.GetHarmonyMethod(nameof(FarmPatcher.After_CleanupBeforePlayerExit))
+            );
+            harmony.Patch(
+                original: this.RequireMethod<GameLocation>(nameof(GameLocation.getRandomTile)),
+                prefix: this.GetHarmonyMethod(nameof(FarmPatcher.After_GetRandomTile))
             );
         }
 
@@ -146,14 +134,6 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
             // change background track
             if (FarmPatcher.ShouldUseBeachMusic())
                 Game1.changeMusicTrack("ocean", music_context: Game1.MusicContext.SubLocation);
-
-            // fix island totem warp
-            // TODO: remove in Stardew Valley 1.5.5
-            if (FarmPatcher.IsOnIslandWarpArrivalTile())
-            {
-                var tile = __instance.GetMapPropertyPosition("WarpTotemEntry", 48, 7);
-                Game1.player.Position = new Vector2(tile.X * Game1.tileSize, tile.Y * Game1.tileSize);
-            }
         }
 
         /// <summary>A method called via Harmony after <see cref="Farm.resetSharedState"/>.</summary>
@@ -166,7 +146,7 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
 
             // toggle campfire (derived from StardewValley.Locations.Mountain:resetSharedState
             Vector2 campfireTile = new Vector2(64, 22);
-            if (FarmPatcher.AddCampfire)
+            if (FarmPatcher.Config.AddCampfire)
             {
                 if (!__instance.objects.ContainsKey(campfireTile))
                 {
@@ -194,22 +174,25 @@ namespace Pathoschild.Stardew.SmallBeachFarm.Patches
                 Game1.changeMusicTrack("none", music_context: Game1.MusicContext.SubLocation);
         }
 
+        /// <summary>A method called via Harmony after <see cref="GameLocation.getRandomTile"/>.</summary>
+        /// <param name="__instance">The farm instance.</param>
+        [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "The naming convention is defined by Harmony.")]
+        private static void After_GetRandomTile(GameLocation __instance, ref Vector2 __result)
+        {
+            if (!FarmPatcher.IsSmallBeachFarm(__instance))
+                return;
+
+            // reduce chance of ocean tiles in random tile selection, which makes things like beach crates much less likely than vanilla
+            Farm farm = (Farm)__instance;
+            int maxTileY = FarmPatcher.Config.EnableIslands ? farm.Map.Layers[0].LayerHeight : 31;
+            for (int i = 0; FarmPatcher.GetFishType(farm, (int)__result.X, (int)__result.Y) == FishType.Ocean && i < 250; i++)
+                __result = new Vector2(Game1.random.Next(farm.Map.Layers[0].LayerWidth), Game1.random.Next(maxTileY));
+        }
+
         /// <summary>Get whether the Small Beach Farm's music should be overridden with the beach sounds.</summary>
         private static bool ShouldUseBeachMusic()
         {
-            return FarmPatcher.UseBeachMusic && !Game1.isRaining;
-        }
-
-        /// <summary>Whether the player is on the arrival tile for the <see cref="IslandWest"/> obelisk warp.</summary>
-        [Obsolete("This will no longer be needed in Stardew Valley 1.5.5.")]
-        private static bool IsOnIslandWarpArrivalTile()
-        {
-            Point curTile = Utility.Vector2ToPoint(Game1.player.Position / Game1.tileSize);
-
-            if (!FarmPatcher.IslandWarpTargets.TryGetValue(Game1.whichFarm, out Point arrivalTile))
-                arrivalTile = FarmPatcher.IslandWarpTargets[-1];
-
-            return curTile == arrivalTile;
+            return FarmPatcher.Config.UseBeachMusic && !Game1.isRaining;
         }
     }
 }