+++ DataLayers/docs/release-notes.md
@@ -1,6 +1,11 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Updated for Stardew Valley 1.5.5, including...
+  * custom scarecrows using the new `obj.IsScarecrow()` and `obj.GetRadiusForScarecrow()` methods;
+  * new `TouchAction Warp` tile property.
+
 ## 1.14.8
 Released 27 November 2021 for SMAPI 3.12.5 or later.